Fixed required keys not impossible to retrieve correctly
A bug with the new `offline-sign-key` parameter was causing the custom
`GetRequiredKeys` handler to set making it impossible to correctly sign
transaction meant to be pushed to the chain.

The problem was the the parameter has been converted from a string to
a string slice and the default value being the empty string slice.

The signing process was incorrectly checking for the parameter being nil
instead of checking for being the empty slice.

This is now fixed.