Added flag `-f` (`--force-unique`) on `tx create`
By using this flag, you ensure the created transaction will be unique.
This work by creating a `eosio.null:nonce` action with the value
being a generate random number.

The nonce is generated using 48 bits entropy read from the random stream.
This should be enough for this use cases of generating unique transaction.