Added possibility to use a plain string for `eosc tx push`
The feature is heuristics based, if the first argument is an existing
file, it is used right away. Otherwise, if it looks like a JSON string
(really basic checks enough to avoid ambiguity with a plain filename), then
the input is considered a JSON transaction. Otherwise, an error is thrown.

Fixes #126