Fixed bad handling of core symbol
The core symbol is now retrieved on-demand when a command requires it. The
current retrieval way is to query the `rammarket` table which should have
a row only when `eosio:init` has been called with the core symbol set.

This row has a quote balance asset value that is always equalt to the
core symbol. This is where we get it.

In the eventuality that it's not possible to retrieve, the user can still
override the default 4,EOS by passing `--core-symbol` to the `eosc` CLI.
This will use this core symbol only when impossible to infer it.

Also added logging facilities to help debugging certain part of the `eosc`
CLI.

Fixes #136